api-server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
psycopg2-binary>=2.9.11
python-dotenv>=1.2.1
pydantic>=2.12.5